Major overhaul of internal RMI stream internals
Now there are no references kept to classes, fields, methods, or
classloaders, not even in caches. This ensures proper behaviour when the
classloaders are unregistered from classloader resolvers. However, it
comes with the cost that these objects need to be looked up every time a
method is called.

However, the lookup compared to the network transfer cost may be
insignificant.

More tests should be made for various use-cases with the RMI to ensure
that we hadn't broken crucial behaviours.