Exclude inaccessible interfaces from proxies
The inaccessible superinterface detection was incomplete. It only
exluded package private interfaces, but didn't filter out interfaces
that are inaccessible through the module system.
This detection is now done by using the
MethodHandles.Lookup.accessibleClass() function that was introduced in
Java 9.

Inaccessible interfaces are also now collected in the RMIStatistics for
debugging and analysis.