Add function to call a remote method on a context variable
This is to allow calling a function on a context variable without first
getting the variable itself. If the caller only needs to call a
function, then this will only require a single request to be made to the
other endpoint, and improve response time.
New protocol commands are added to protocol version 2, the added
function is backward compatible.