Change RMIVariables.closeWait to RMIVariables.waitClosure
The closing and waiting for the closure should be separate methods, so
it can be better controlled by the user. Furthermore, the named RMI
variables perform reference counting, making the close() method
non-recallable, therefore would make closeWait() non-recallable too.
However, the waiting should be recallable, because it may get
interrupted.